% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_ggplot.R
\name{notify_ggplot}
\alias{notify_ggplot}
\title{Send a ggplot object to LINE}
\usage{
notify_ggplot(message = "", token = Sys.getenv("LINE_API_TOKEN"),
  plot = last_plot(), path = NULL, scale = 1, width = par("din")[1],
  height = par("din")[2], units = c("in", "cm", "mm"), dpi = 300,
  limitsize = TRUE, ...)
}
\arguments{
\item{message}{message (chr)}

\item{token}{the LINE Notify personal access token (chr)}

\item{plot}{ggplot object to save, defaults to last plot displayed}

\item{path}{path to save plot. (chr)}

\item{scale}{scaling factor}

\item{width}{width (defaults to the width of current plotting window)}

\item{height}{height (defaults to the height of current plotting window)}

\item{units}{units for width and height when either one is explicitly specified
(in, cm, or mm)}

\item{dpi}{dpi to use for raster graphics}

\item{limitsize}{when TRUE (the default), ggsave will not save images larger
than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{...}{other arguments passed to graphics device}
}
\value{
\code{httr} response object (invisibly)
}
\description{
Send a ggplot object to LINE
}
\examples{
\dontrun{
library(ggplot2)
Sys.setenv(LINE_API_TOKEN="xxxxxxxxxxxxxxxxxx")
data(iris)
p <- ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
 geom_point()
notify_ggplot("Plotting the Iris", plot = p)
}
}
