% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_msg.R
\name{notify_msg}
\alias{notify_msg}
\title{Send a message to LINE}
\usage{
notify_msg(message, token = Sys.getenv("LINE_API_TOKEN"))
}
\arguments{
\item{message}{message (chr)}

\item{token}{the LINE Notify full API token (chr)}
}
\value{
\code{httr} response object (invisibly)
}
\description{
Send a message to LINE
}
\examples{
\dontrun{
Sys.setenv(LINE_API_TOKEN="xxxxxxxxxxxxxxxxxx")
notify_msg("hello world")
}
}
