% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_plot.R
\name{notify_plot}
\alias{notify_plot}
\title{Send a plot to LINE}
\usage{
notify_plot(message = "", token = Sys.getenv("LINE_API_TOKEN"),
  file = "line_notify")
}
\arguments{
\item{message}{message (chr)}

\item{token}{the LINE Notify personal access token (chr)}

\item{file}{prefix for filenames}
}
\value{
\code{httr} response object (invisibly)
}
\description{
Send a plot to LINE
}
\examples{
\dontrun{
Sys.setenv(LINE_API_TOKEN="xxxxxxxxxxxxxxxxxx")
plot(iris)
notify_plot("Plotting the Iris")
}
}
