% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_ggplot.R
\name{rand_strings}
\alias{rand_strings}
\title{generate random strings}
\usage{
rand_strings(len = 16)
}
\arguments{
\item{len}{length of random strings}
}
\value{
random strings
}
\description{
generate random strings
}
